Extract FrameTimer to common module
