+++ src/common/mod.rs
@@ -1,3 +1,6 @@
+mod frame_timer;
+pub use self::frame_timer::*;
+
 #[allow(dead_code)]
 pub mod fonts {
     pub const SPACE_MONO_REGULAR: &str = "SpaceMono-Regular";
